// src/infrastructure/grpc/generated/course/types/module.ts
export interface UpdateModuleRequest {
  moduleId: string;
  userId: string;
  courseId: string;
  title: string;
  description: string;
  isPublished: boolean;
  order: number;
}

export interface DeleteModuleRequest {
  moduleId: string;
  courseId: string;
  userId: string;
}

export interface GetModulesByCourseRequest {
  courseId: string;
}

export interface ModuleData {
  id: string;
  courseId: string;
  title: string;
  lessons: LessonData[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  description: string;
  isPublished: boolean;
  order: number;
  quiz: QuizData | undefined;
}

export interface ModuleResponse {
  module?: ModuleData | undefined;
  error?: Error | undefined;
}

export interface ModulesData {
  modules: ModuleData[];
}

export interface ModulesResponse {
  modules?: ModulesData | undefined;
  error?: Error | undefined;
}

export interface DeleteModuleResponse {
  success?: DeleteSuccess | undefined;
  error?: Error | undefined;
}
// ...
export const GetModulesByCourseRequest: MessageFns<GetModulesByCourseRequest> =
  {
    encode(
      message: GetModulesByCourseRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.courseId !== "") {
        writer.uint32(10).string(message.courseId);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetModulesByCourseRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetModulesByCourseRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.courseId = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetModulesByCourseRequest {
      return {
        courseId: isSet(object.courseId)
          ? globalThis.String(object.courseId)
          : isSet(object.course_id)
            ? globalThis.String(object.course_id)
            : "",
      };
    },

    toJSON(message: GetModulesByCourseRequest): unknown {
      const obj: any = {};
      if (message.courseId !== "") {
        obj.courseId = message.courseId;
      }
      return obj;
    },

    create<I extends Exact<DeepPartial<GetModulesByCourseRequest>, I>>(
      base?: I,
    ): GetModulesByCourseRequest {
      return GetModulesByCourseRequest.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<GetModulesByCourseRequest>, I>>(
      object: I,
    ): GetModulesByCourseRequest {
      const message = createBaseGetModulesByCourseRequest();
      message.courseId = object.courseId ?? "";
      return message;
    },
  };
// ...
export const ModuleResponse: MessageFns<ModuleResponse> = {
  encode(
    message: ModuleResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.module !== undefined) {
      ModuleData.encode(message.module, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ModuleResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseModuleResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.module = ModuleData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ModuleResponse {
    return {
      module: isSet(object.module)
        ? ModuleData.fromJSON(object.module)
        : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: ModuleResponse): unknown {
    const obj: any = {};
    if (message.module !== undefined) {
      obj.module = ModuleData.toJSON(message.module);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ModuleResponse>, I>>(
    base?: I,
  ): ModuleResponse {
    return ModuleResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ModuleResponse>, I>>(
    object: I,
  ): ModuleResponse {
    const message = createBaseModuleResponse();
    message.module =
      object.module !== undefined && object.module !== null
        ? ModuleData.fromPartial(object.module)
        : undefined;
    message.error =
      object.error !== undefined && object.error !== null
        ? Error.fromPartial(object.error)
        : undefined;
    return message;
  },
};
// ...
export const ModulesResponse: MessageFns<ModulesResponse> = {
  encode(
    message: ModulesResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.modules !== undefined) {
      ModulesData.encode(message.modules, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ModulesResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseModulesResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.modules = ModulesData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ModulesResponse {
    return {
      modules: isSet(object.modules)
        ? ModulesData.fromJSON(object.modules)
        : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: ModulesResponse): unknown {
    const obj: any = {};
    if (message.modules !== undefined) {
      obj.modules = ModulesData.toJSON(message.modules);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ModulesResponse>, I>>(
    base?: I,
  ): ModulesResponse {
    return ModulesResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ModulesResponse>, I>>(
    object: I,
  ): ModulesResponse {
    const message = createBaseModulesResponse();
    message.modules =
      object.modules !== undefined && object.modules !== null
        ? ModulesData.fromPartial(object.modules)
        : undefined;
    message.error =
      object.error !== undefined && object.error !== null
        ? Error.fromPartial(object.error)
        : undefined;
    return message;
  },
};

function createBaseDeleteModuleResponse(): DeleteModuleResponse {
  return { success: undefined, error: undefined };
}

export const DeleteModuleResponse: MessageFns<DeleteModuleResponse> = {
  encode(
    message: DeleteModuleResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.success !== undefined) {
      DeleteSuccess.encode(message.success, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): DeleteModuleResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteModuleResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.success = DeleteSuccess.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteModuleResponse {
    return {
      success: isSet(object.success)
        ? DeleteSuccess.fromJSON(object.success)
        : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: DeleteModuleResponse): unknown {
    const obj: any = {};
    if (message.success !== undefined) {
      obj.success = DeleteSuccess.toJSON(message.success);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteModuleResponse>, I>>(
    base?: I,
  ): DeleteModuleResponse {
    return DeleteModuleResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteModuleResponse>, I>>(
    object: I,
  ): DeleteModuleResponse {
    const message = createBaseDeleteModuleResponse();
    message.success =
      object.success !== undefined && object.success !== null
        ? DeleteSuccess.fromPartial(object.success)
        : undefined;
    message.error =
      object.error !== undefined && object.error !== null
        ? Error.fromPartial(object.error)
        : undefined;
    return message;
  },
};

type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}